timerandom.gawk: add gawk-finding magic and timing
This works around the one-argument "#!" interpreter limitation so that
multiple arguments can be passed to the AWK interpreter while at the
same time searching for a non-standard named variant in $PATH.

To do this we use an opening statement that both /bin/sh and AWKs can
parse, but which does nothing in AWK, and set the "#!" interpreter to
/bin/sh, which evals and execs a command line for "env" that can
find (by searching in $PATH) and start the AWK interpreter with the
necessary '-f' option, and pass the remaining command line parameters to
it.

Also add simple timekeeping to see roughly how long the run takes.