t-named-params.c:  potential dangers of struct literals for named parameters
Add an example showing how struct literal storage is ephemeral and might
be reused if the same type of literal is initialised several times or
in multiple places in the code.

Show it is safer to copy literals (or pass by value) instead of keeping
references to their storage location.